vggt/requirements_demo.txt
gradio==5.17.1
viser==0.2.23
tqdm
hydra-core
omegaconf
opencv-python
scipy
onnxruntime
requests
trimesh
matplotlib
# feel free to skip the dependencies below if you do not need demo_colmap.py
pycolmap==3.10.0
pyceres==2.3
git+https://github.com/jytime/LightGlue.git#egg=lightglue

